Refactor folder structure.
