+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import './App.css';
-import Navbar from './components/navbar';
+import Navbar from './components/navbar/navbar';
 import Missions from './pages/missions';
 import Profile from './pages/profile';
 import Rockets from './pages/rockets';
